requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.0
    # via django
astroid==2.9.3
    # via pylint
attrs==21.4.0
    # via pytest
black==22.1.0
    # via -r requirements.in
click==8.0.4
    # via black
django==4.0.3
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   djangorestframework
    #   drf-nested-routers
django-cors-headers==3.11.0
    # via -r requirements.in
django-extensions==3.1.5
    # via -r requirements.in
django-filter==21.1
    # via -r requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   drf-nested-routers
drf-nested-routers==0.93.4
    # via -r requirements.in
factory-boy==3.2.1
    # via -r requirements.in
faker==13.3.1
    # via factory-boy
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via pylint
lazy-object-proxy==1.7.1
    # via astroid
mccabe==0.6.1
    # via pylint
mypy-extensions==0.4.3
    # via black
packaging==21.3
    # via pytest
pathspec==0.9.0
    # via black
pep8==1.7.1
    # via -r requirements.in
platformdirs==2.5.1
    # via
    #   black
    #   pylint
pluggy==1.0.0
    # via pytest
psycopg2==2.9.3
    # via -r requirements.in
py==1.11.0
    # via pytest
pylint==2.12.2
    # via -r requirements.in
pyparsing==3.0.7
    # via packaging
pytest==7.1.0
    # via
    #   -r requirements.in
    #   pytest-django
pytest-django==4.5.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via faker
pytz==2021.3
    # via djangorestframework
six==1.16.0
    # via python-dateutil
sqlparse==0.4.2
    # via django
toml==0.10.2
    # via pylint
tomli==2.0.1
    # via
    #   black
    #   pytest
wrapt==1.13.3
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
